# swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HubNumm",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "https://p.hubnumm.pprv.eu",
            "description": "HubNumm"
        }
    ],
    "paths": {
...
        "/api/stats/kerry": {
            "post": {
                "tags": [
                    "Statistiques"
                ],
                "summary": "Returns a list of gross sales statistics for Kerry",
                "operationId": "65af759acbbc4251d45c51e10c93875f",
                "responses": {
                    "200": {
                        "description": "Liste de CA triés par type et par conseillers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Réponse",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "type": "object",
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/StatisticClientRivalis"
                                                },
                                                {
                                                    "$ref": "#/components/schemas/StatisticCMCIC"
                                                },
                                                {
                                                    "$ref": "#/components/schemas/StatisticComnat"
                                                },
                                                {
                                                    "$ref": "#/components/schemas/StatisticLeasis"
                                                },
                                                {
                                                    "$ref": "#/components/schemas/StatisticLicenceDirecte"
                                                },
                                                {
                                                    "$ref": "#/components/schemas/StatisticProgrammeCroissance"
                                                },
                                                {
                                                    "$ref": "#/components/schemas/StatisticRedevance"
                                                },
                                                {
                                                    "$ref": "#/components/schemas/StatisticRenouvellement"
                                                },
                                                {
                                                    "$ref": "#/components/schemas/StatisticRivacentrale"
                                                },
                                                {
                                                    "$ref": "#/components/schemas/StatisticRivashop"
                                                },
                                                {
                                                    "$ref": "#/components/schemas/StatisticTotalReseau"
                                                },
                                                {
                                                    "$ref": "#/components/schemas/StatisticTotalTPE"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            }
        },
        "/stats/kerry-withoutid": {
            "post": {
                "tags": [
                    "Statistiques"
                ],
                "summary": "Returns a list of gross sales statistics for Kerry",
                "operationId": "f5c5fdff2db93cf72894d3e9ff8e1f1e",
                "responses": {
                    "200": {
                        "description": "Liste de CA triés par type et par conseillers dans le cas où le conseiller n'a pas d'ID Datalake renseigné dans Numm",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Statistiques sans ID Datalake",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/StatisticWithoutIDClientRivalis"
                                                },
                                                {
                                                    "$ref": "#/components/schemas/StatisticWithoutIDCMCIC"
                                                },
                                                {
                                                    "$ref": "#/components/schemas/StatisticWithoutIDComnat"
                                                },
                                                {
                                                    "$ref": "#/components/schemas/StatisticWithoutIDLeasis"
                                                },
                                                {
                                                    "$ref": "#/components/schemas/StatisticWithoutIDLicenceDirecte"
                                                },
                                                {
                                                    "$ref": "#/components/schemas/StatisticWithoutIDProgrammeCroissance"
                                                },
                                                {
                                                    "$ref": "#/components/schemas/StatisticWithoutIDRedevance"
                                                },
                                                {
                                                    "$ref": "#/components/schemas/StatisticWithoutIDRenouvellement"
                                                },
                                                {
                                                    "$ref": "#/components/schemas/StatisticWithoutIDRivacentrale"
                                                },
                                                {
                                                    "$ref": "#/components/schemas/StatisticWithoutIDRivashop"
                                                },
                                                {
                                                    "$ref": "#/components/schemas/StatisticWithoutIDTotalReseau"
                                                },
                                                {
                                                    "$ref": "#/components/schemas/StatisticWithoutIDTotalTPE"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            }
        },
        "/stats/analyticCompta": {
            "post": {
                "tags": [
                    "Statistiques",
                    "Analytique",
                    "BU"
                ],
                "summary": "Returns a accounting data by analytics",
                "operationId": "7b1f8ba2ddd2daa84bee694b39b9ce8d",
                "responses": {
                    "200": {
                        "description": "Liste de données analytiques par BU et par catégorie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Données de la comptabilité analytique, par BU et catégorie",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/ExternalExpenses"
                                                },
                                                {
                                                    "$ref": "#/components/schemas/Outsourcing"
                                                },
                                                {
                                                    "$ref": "#/components/schemas/Sales"
                                                },
                                                {
                                                    "$ref": "#/components/schemas/SocialExpenses"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "StatisticCMCIC": {
                "properties": {
                    "CA_CM_CIC": {
                        "properties": {
                            "ID_DATALAKE_DU_CONSEILLER": {
                                "type": "number",
                                "format": "float",
                                "example": 123.01
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "StatisticWithoutIDCMCIC": {
                "properties": {
                    "NOM_DU_CONSEILLER": {
                        "properties": {
                            "CA_CM_CIC": {
                                "type": "number",
                                "format": "float",
                                "example": 12.2
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "StatisticClientRivalis": {
                "properties": {
                    "CA_CLIENT_RIVALIS": {
                        "properties": {
                            "ID_DATALAKE_DU_CONSEILLER": {
                                "type": "number",
                                "format": "float",
                                "example": 123.01
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "StatisticWithoutIDClientRivalis": {
                "properties": {
                    "NOM_DU_CONSEILLER": {
                        "properties": {
                            "CA_CLIENT_RIVALIS": {
                                "type": "number",
                                "format": "float",
                                "example": 12.2
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "StatisticComnat": {
                "properties": {
                    "CA_COMNAT": {
                        "properties": {
                            "ID_DATALAKE_DU_CONSEILLER": {
                                "type": "number",
                                "format": "float",
                                "example": 123.01
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "StatisticWithoutIDComnat": {
                "properties": {
                    "NOM_DU_CONSEILLER": {
                        "properties": {
                            "CA_COMNAT": {
                                "type": "number",
                                "format": "float",
                                "example": 12.2
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "StatisticLeasis": {
                "properties": {
                    "CA_LEASIS": {
                        "properties": {
                            "ID_DATALAKE_DU_CONSEILLER": {
                                "type": "number",
                                "format": "float",
                                "example": 123.01
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "StatisticWithoutIDLeasis": {
                "properties": {
                    "NOM_DU_CONSEILLER": {
                        "properties": {
                            "CA_LEASIS": {
                                "type": "number",
                                "format": "float",
                                "example": 12.2
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "StatisticLicenceDirecte": {
                "properties": {
                    "CA_LICENCE_DIRECTE": {
                        "properties": {
                            "ID_DATALAKE_DU_CONSEILLER": {
                                "type": "number",
                                "format": "float",
                                "example": 123.01
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "StatisticProgrammeCroissance": {
                "properties": {
                    "CA_PROGRAMME_CROISSANCE": {
                        "properties": {
                            "ID_DATALAKE_DU_CONSEILLER": {
                                "type": "number",
                                "format": "float",
                                "example": 123.01
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "StatisticRedevance": {
                "properties": {
                    "CA_REDEVANCE": {
                        "properties": {
                            "ID_DATALAKE_DU_CONSEILLER": {
                                "type": "number",
                                "format": "float",
                                "example": 123.01
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "StatisticRenouvellement": {
                "properties": {
                    "CA_RENOUVELLEMENT": {
                        "properties": {
                            "ID_DATALAKE_DU_CONSEILLER": {
                                "type": "number",
                                "format": "float",
                                "example": 123.01
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "StatisticRivacentrale": {
                "properties": {
                    "CA_RIVACENTRALE": {
                        "properties": {
                            "ID_DATALAKE_DU_CONSEILLER": {
                                "type": "number",
                                "format": "float",
                                "example": 123.01
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "StatisticRivashop": {
                "properties": {
                    "CA_RIVASHOP": {
                        "properties": {
                            "ID_DATALAKE_DU_CONSEILLER": {
                                "type": "number",
                                "format": "float",
                                "example": 123.01
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "StatisticTotalReseau": {
                "properties": {
                    "CA_TOTAL_RESEAU": {
                        "properties": {
                            "ID_DATALAKE_DU_CONSEILLER": {
                                "type": "number",
                                "format": "float",
                                "example": 123.01
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "StatisticTotalTPE": {
                "properties": {
                    "CA_TOTAL_TPE": {
                        "properties": {
                            "ID_DATALAKE_DU_CONSEILLER": {
                                "type": "number",
                                "format": "float",
                                "example": 123.01
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "bearerFormat": "JWT"
            }
        }
    }
}